Implement Error Handling for Time Units

+++ RecurringTask.java
@@ -5,7 +5,7 @@ class RecurringTask extends Task {
     // String[] recurringTaskTypes = {"Class", "Study", "Sleep", "Exercise", "Work", "Meal"};
 
     // Constructor to initialize recurring task properties
-    public RecurringTask(String id, String name, String startTime, String type, int duration, String startDate, String endDate) {
+    public RecurringTask(String id, String name, String startTime, String type, double duration, String startDate, String endDate) {
         super(id, name, startTime, type, duration);
         this.startDate = startDate;
         this.endDate = endDate;
